refactor(world): de-part fog/capital/army libs; move topology identity (#3968)
Convert remaining world lib part fragments to explicit-import libraries,
extract topology identity from naval.dart, add repo.world_no_part_directives,
and wire spyRevealFogGame into fog spy-decay pins.
