+++ SPEC/program/repo-lint.md
@@ -113,6 +113,7 @@
 | `tool/check_combat_test_duplicate_descriptions.dart` | Preserve unique `test('…')` descriptions during combat table merges — fail when the same description appears in more than one file under `packages/colonizethis_combat/test/**` (Refs #3865); rule `repo.combat_test_duplicate_descriptions` |
 | `tool/check_orders_no_cross_package_src_imports.dart` | Forbid `colonizethis_orders/lib/**` from deep-importing another colonizethis package's private `lib/src/**` tree: any `import 'package:colonizethis_<pkg>/src/...'` directive where `<pkg>` is not `orders` is a violation; cross-package symbols must be consumed through the owning package's public barrel `package:colonizethis_<pkg>/colonizethis_<pkg>.dart`. Self-imports of `package:colonizethis_orders/src/...` and relative imports are out of scope; generated suffixes (`*.g.dart`, `*.freezed.dart`, `*.mocks.dart`) are skipped. Stronger than `repo.domain_package_barrel_import` (which only blocks deep imports of barrel-*published* files): this rule blocks **all** sibling `src/` imports so the orders package never reaches into another package's internals (Refs #3543); rule `repo.orders_no_cross_package_src_imports` — see `SPEC/program/logic-package-barrel-contracts.md` |
 | `tool/check_world_no_circular_imports.dart` | Build the file-level `import` / `export` graph among `packages/colonizethis_world/lib/**` hand-written Dart files — resolving relative imports and `package:colonizethis_world/...` self-imports to files inside that tree, ignoring `dart:` / other-package URIs, `part` / `part of` directives, and generated suffixes (`*.g.dart`, `*.freezed.dart`, `*.mocks.dart`) — and fail when any directed cycle exists. Complements `repo.world_no_logic_deps` (cross-package logic deps) and `repo.domain_package_import_dag` (cross-package direction) by catching intra-package cycles such as the former `province_lookup.dart` ↔ `province_owner_cache.dart` and `naval.dart` ↔ `topology_helpers.dart` two-node cycles, which prevented either file from being understood or unit-tested in isolation. The remediation is to extract the shared symbol into a third file imported by both sides; the checker reports one offending cycle as an ordered file chain (Refs #3544); rule `repo.world_no_circular_imports` |
+| `tool/check_world_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_world/lib/**` so fog, capital/GP fall, and army-migration sub-files stay proper libraries with explicit imports (mirrors turn/diplomacy/orders no-part gates; Refs #3968); rule `repo.world_no_part_directives` |
 | `tool/check_logic_dual_region_province_field_access.dart` | Caps direct `oldWorld/newWorld` `provinces` / `units` references in `packages/colonizethis_world/lib/src/**` outside canonical lookup files (`province_lookup.dart`, `unit_lookup.dart`) (GitHub #2071; post-split scan root Refs #3290); rule `repo.logic_dual_region_province_field_access` — see `SPEC/program/logic-dual-region-province-access.md` |
 | `tool/check_world_no_duplicate_extension_helper.dart` | Forbid a top-level helper in `packages/colonizethis_world/lib/**` from re-implementing a same-named `WorldState` extension method instead of delegating to it. AST-based: any top-level function whose name is in the watched set (initially `allProvinces`) must have a thin delegating body — either `=> <target>.<name>()` or a single `{ return <target>.<name>(); }` calling the same-named no-argument extension method — otherwise it is flagged as a parallel definition. Closes the gap that let the duplicate `allProvinces(WorldState)` free function and `WorldState.allProvinces()` extension drift apart (Refs #3710); rule `repo.world_no_duplicate_extension_helper` |
 | `tool/check_logic_units_by_id_rebuild.dart` | Forbid non-canonical `unitsByIdFromWorld(` call sites across the split domain package source trees (`world`, `combat`, `economy`, `diplomacy`, `setup`, `orders`, `turn`, `ai_contracts` `lib/src/**`, plus the thin `colonizethis_logic/lib/src`), excluding the canonical `packages/colonizethis_world/lib/src/world/unit_lookup.dart`; consumers MUST use the cached `WorldState.allUnitsById` getter so the per-`WorldState` unit index is reused on the order-suggestion / turn-resolution hot path (GitHub #2836 AC 2; post-split scan roots Refs #3290); rule `repo.logic_units_by_id_rebuild` |
@@ -369,6 +370,9 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given the `colonizethis_world` package on `dev` where no `lib/**` Dart file participates in an `import` / `export` cycle with another file in that tree, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_no_circular_imports`, then the rule reports the scanned file count, passes without violations, and exits `0` (Refs #3544).
 - Given a `packages/colonizethis_world/lib/**` Dart file that imports a second world `lib/**` file which (directly or transitively) imports the first back, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_no_circular_imports`, then the run fails listing the offending cycle as an ordered file chain and exits `1`, directing the author to break the cycle by extracting the shared symbol into a third file (Refs #3544).
 - Given the `packages/colonizethis_world/lib` tree is absent, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_no_circular_imports`, then the rule reports the missing tree and exits `1`.
+
+- Given the `colonizethis_world` package on `dev` where no `lib/**` file uses a `part` / `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_no_part_directives`, then the rule passes without violations and exits `0` (Refs #3968).
+- Given a `packages/colonizethis_world/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #3968).
 - Given the `colonizethis_world` package on `dev` where the free `allProvinces(WorldState)` function delegates to the `WorldState.allProvinces()` extension method, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_no_duplicate_extension_helper`, then the rule passes without violations and exits `0` (Refs #3710).
 - Given a top-level function under `packages/colonizethis_world/lib/**` named `allProvinces` whose body re-implements dual-region traversal (for example `sync* { yield* world.oldWorld.provinces; yield* world.newWorld.provinces; }`) instead of delegating to `world.allProvinces()`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.world_no_duplicate_extension_helper`, then the run fails listing each offending `path:line` and exits `1`, directing the author to delegate to the `WorldState` extension method (Refs #3710).
 - Given the `app/lib` tree on `dev` where only `app/lib/features/game/widgets/shell/shell_player_context.dart` (the declaration site) and `app/lib/features/game/widgets/shell/shell_player_guarded_body.dart` (the canonical helper) reference `shellPanelsNotDefined`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.app_no_shell_panel_duplication`, then the rule passes without violations and exits `0` (Refs #3546).
