+++ tool/ct_repo_lint_manifest.yaml
@@ -635,6 +635,13 @@ rules:
     runner: dart
     script: tool/check_orders_no_part_directives.dart
 
+  - rule_id: repo.world_no_part_directives
+    group: structure
+    title: colonizethis_world lib must not use Dart part / part of directives (Refs #3968)
+    spec: SPEC/program/world-no-part-directives.md
+    runner: dart
+    script: tool/check_world_no_part_directives.dart
+
   - rule_id: repo.no_flame_in_widgets
     group: structure
     title: app/lib/widgets must not import package:flame directly
